Move semantic tokens provider to the relevant file
