+++ src/extension.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
-import { spawnSync } from 'node:child_process';
 import { output } from './logging';
-import { type TokenModifier, type TokenType, legend } from './semanticTokens';
+import { legend, DocumentSemanticTokensProvider } from './semanticTokens';
 
 // Inspired by https://github.com/microsoft/vscode-extension-samples/tree/main/semantic-tokens-sample
 
@@ -14,60 +13,4 @@ export function activate(context: vscode.ExtensionContext) {
       legend
     )
   );
-}
-
-interface ParsedToken {
-  line: number;
-  startCharacter: number;
-  length: number;
-  tokenType: TokenType;
-  tokenModifiers: TokenModifier[];
-}
-
-class DocumentSemanticTokensProvider
-  implements vscode.DocumentSemanticTokensProvider
-{
-  async provideDocumentSemanticTokens(
-    document: vscode.TextDocument,
-    token: vscode.CancellationToken
-  ): Promise<vscode.SemanticTokens> {
-    // output.appendLine(`Parsing file "${document.uri}"`);
-    const allTokens: ParsedToken[] = this._parseText(document.getText());
-    const builder = new vscode.SemanticTokensBuilder(legend);
-    allTokens.forEach((token) => {
-      builder.push(
-        new vscode.Range(
-          new vscode.Position(token.line, token.startCharacter),
-          new vscode.Position(token.line, token.startCharacter + token.length)
-        ),
-        token.tokenType,
-        token.tokenModifiers
-      );
-    });
-    return builder.build();
-  }
-
-  private _parseText(doc: string): ParsedToken[] {
-    const path =
-      vscode.workspace.getConfiguration().get<string | null>('rzk.path') ??
-      'rzk';
-    const processResult = spawnSync(path, ['tokenize'], { input: doc });
-    if (processResult.error) {
-      const { message, stack } = processResult.error;
-      output.appendLine('Error running rzk:' + message + '\n' + stack);
-      return [];
-    }
-    if (processResult.stderr.length) {
-      output.appendLine(
-        'Error tokenizing:\n' + processResult.stderr.toString()
-      );
-      return [];
-    }
-    const stdout = processResult.stdout.toString();
-    try {
-      return JSON.parse(stdout);
-    } catch {
-      return [];
-    }
-  }
 }
