+++ src/semanticTokens.ts
@@ -1,4 +1,6 @@
 import * as vscode from 'vscode';
+import { spawnSync } from 'node:child_process';
+import { output } from './logging';
 
 // https://code.visualstudio.com/api/language-extensions/semantic-highlight-guide#standard-token-types-and-modifiers
 export const tokenTypes = [
@@ -49,3 +51,59 @@ export const legend = new vscode.SemanticTokensLegend(
 
 export type TokenType = (typeof tokenTypes)[number];
 export type TokenModifier = (typeof tokenModifiers)[number];
+
+interface ParsedToken {
+  line: number;
+  startCharacter: number;
+  length: number;
+  tokenType: TokenType;
+  tokenModifiers: TokenModifier[];
+}
+
+export class DocumentSemanticTokensProvider
+  implements vscode.DocumentSemanticTokensProvider
+{
+  async provideDocumentSemanticTokens(
+    document: vscode.TextDocument,
+    token: vscode.CancellationToken
+  ): Promise<vscode.SemanticTokens> {
+    // output.appendLine(`Parsing file "${document.uri}"`);
+    const allTokens: ParsedToken[] = this._parseText(document.getText());
+    const builder = new vscode.SemanticTokensBuilder(legend);
+    allTokens.forEach((token) => {
+      builder.push(
+        new vscode.Range(
+          new vscode.Position(token.line, token.startCharacter),
+          new vscode.Position(token.line, token.startCharacter + token.length)
+        ),
+        token.tokenType,
+        token.tokenModifiers
+      );
+    });
+    return builder.build();
+  }
+
+  private _parseText(doc: string): ParsedToken[] {
+    const path =
+      vscode.workspace.getConfiguration().get<string | null>('rzk.path') ??
+      'rzk';
+    const processResult = spawnSync(path, ['tokenize'], { input: doc });
+    if (processResult.error) {
+      const { message, stack } = processResult.error;
+      output.appendLine('Error running rzk:' + message + '\n' + stack);
+      return [];
+    }
+    if (processResult.stderr.length) {
+      output.appendLine(
+        'Error tokenizing:\n' + processResult.stderr.toString()
+      );
+      return [];
+    }
+    const stdout = processResult.stdout.toString();
+    try {
+      return JSON.parse(stdout);
+    } catch {
+      return [];
+    }
+  }
+}
